Restructure base folder, add src and environments folders
