+++ base/EnvironmentStateValidityChecker.h
@@ -1,0 +1,3 @@
+#include <ompl/base/StateValidityChecker.h>
+
+#include "base/Environment.h"
