+++ experiments/benchmark.cpp
@@ -1,13 +1,11 @@
 #include <chrono>
 #include <ctime>
 
-#include "base/GridMaze.h"
 #include "base/PlannerSettings.h"
-
+#include "base/environments/GridMaze.h"
 #include "planners/OMPLPlanner.hpp"
 #include "planners/sbpl/SbplPlanner.h"
 #include "planners/thetastar/ThetaStar.h"
-
 #include "utils/PathEvaluation.hpp"
 #include "utils/ScenarioLoader.h"
