+++ experiments/benchmark_corridors.cpp
@@ -1,10 +1,8 @@
 #include "base/PlannerSettings.h"
-#include "base/GridMaze.h"
-
+#include "base/environments/GridMaze.h"
 #include "planners/OMPLPlanner.hpp"
 #include "planners/sbpl/SbplPlanner.h"
 #include "planners/thetastar/ThetaStar.h"
-
 #include "utils/PathEvaluation.hpp"
 
 namespace og = ompl::geometric;
