+++ experiments/compute_distances.cpp
@@ -1,9 +1,8 @@
+#include "base/PlannerSettings.h"
+#include "base/environments/GridMaze.h"
 #include "utils/Log.h"
 #include "utils/ScenarioLoader.h"
 #include "utils/Stopwatch.hpp"
-
-#include "base/GridMaze.h"
-#include "base/PlannerSettings.h"
 
 namespace og = ompl::geometric;
