+++ experiments/generate_corridor_radii_benchmarks.cpp
@@ -2,10 +2,9 @@
 #include <iomanip>
 #include <iostream>
 
-#include "utils/Log.h"
-
-#include "base/PolygonMaze.h"
 #include "base/PlannerSettings.h"
+#include "base/environments/PolygonMaze.h"
+#include "utils/Log.h"
 
 namespace og = ompl::geometric;
