+++ experiments/generate_obstacle_ratio_benchmarks.cpp
@@ -3,8 +3,7 @@
 #include <iostream>
 
 #include "base/PlannerSettings.h"
-#include "base/PolygonMaze.h"
-
+#include "base/environments/PolygonMaze.h"
 #include "utils/Log.h"
 #include "utils/PlannerUtils.hpp"
