+++ experiments/generate_turning_radii_benchmarks.cpp
@@ -2,9 +2,8 @@
 #include <iomanip>
 #include <iostream>
 
-#include "base/PolygonMaze.h"
 #include "base/PlannerSettings.h"
-
+#include "base/environments/PolygonMaze.h"
 #include "utils/Log.h"
 
 namespace og = ompl::geometric;
